+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## [Unreleased]
 
+### Added
+- Support for `application/x-www-form-urlencoded`
+- Export `HttpError` class
+
 ## [0.2.1] - 2023-11-16
 ### Added
 - Better support for error responses with `unwrapError`.